+++ CHANGELOG.md
@@ -21,6 +21,11 @@ Things to be included in the next release go here.
 ### Fixed
 
 - Updated the `TekScope.curve_query()` method to properly create a csv file without any trailing commas that previously caused csv parsing issues.
+- Updated `_AbstractDeviceVISAWriteQueryControl.expect_esr()`:
+    - ([#437](https://github.com/tektronix/tm_devices/issues/437)) Now uses \_no_error_string with regex matching by default.
+    - ([#433](https://github.com/tektronix/tm_devices/issues/433)) Better error reporting and properly handle when not enough error messages are returned from the device.
+- Updated `tsp_control.get_buffers()` to use subindex rather than improperly strip characters for buffer_size_name.
+- Updated `tm_devices/helpers/verification_functions.verify_values` helper function to use Decimal to calculate percentage tolerance to avoid floating point math anomalies.
 
 ---